+++ node/src/Commands.ts
@@ -939,6 +939,10 @@ export type ZAddOptions = {
      * is greater than the current score. Equivalent to `GT` in the Redis API.
      */
     updateOptions?: "scoreLessThanCurrent" | "scoreGreaterThanCurrent";
+    /**
+     * Modify the return value from the number of new elements added, to the total number of elements changed.
+     */
+    changed?: boolean;
 };
 
 /**
@@ -948,7 +952,7 @@ export function createZAdd(
     key: string,
     membersScoresMap: Record<string, number>,
     options?: ZAddOptions,
-    changedOrIncr?: "CH" | "INCR",
+    incr: boolean = false,
 ): command_request.Command {
     let args = [key];
 
@@ -970,10 +974,14 @@ export function createZAdd(
         } else if (options.updateOptions === "scoreGreaterThanCurrent") {
             args.push("GT");
         }
+
+        if (options.changed) {
+            args.push("CH");
+        }
     }
 
-    if (changedOrIncr) {
-        args.push(changedOrIncr);
+    if (incr) {
+        args.push("INCR");
     }
 
     args = args.concat(
@@ -1874,8 +1882,63 @@ export function createGeoAdd(
         args = args.concat(coord.toArgs());
         args.push(member);
     });
-
     return createCommand(RequestType.GeoAdd, args);
+}
+
+/**
+ * Enumeration representing distance units options for the {@link geodist} command.
+ */
+export enum GeoUnit {
+    /** Represents distance in meters. */
+    METERS = "m",
+
+    /** Represents distance in kilometers. */
+    KILOMETERS = "km",
+
+    /** Represents distance in miles. */
+    MILES = "mi",
+
+    /** Represents distance in feet. */
+    FEET = "ft",
+}
+
+/**
+ * @internal
+ */
+export function createGeoPos(
+    key: string,
+    members: string[],
+): command_request.Command {
+    return createCommand(RequestType.GeoPos, [key].concat(members));
+}
+
+/**
+ * @internal
+ */
+export function createGeoDist(
+    key: string,
+    member1: string,
+    member2: string,
+    geoUnit?: GeoUnit,
+): command_request.Command {
+    const args: string[] = [key, member1, member2];
+
+    if (geoUnit) {
+        args.push(geoUnit);
+    }
+
+    return createCommand(RequestType.GeoDist, args);
+}
+
+/**
+ * @internal
+ */
+export function createGeoHash(
+    key: string,
+    members: string[],
+): command_request.Command {
+    const args: string[] = [key].concat(members);
+    return createCommand(RequestType.GeoHash, args);
 }
 
 /**
@@ -1896,4 +1959,34 @@ export function createZRevRankWithScore(
     member: string,
 ): command_request.Command {
     return createCommand(RequestType.ZRevRank, [key, member, "WITHSCORE"]);
+}
+
+/**
+ * Mandatory option for zmpop.
+ * Defines which elements to pop from the sorted set.
+ */
+export enum ScoreFilter {
+    /** Pop elements with the highest scores. */
+    MAX = "MAX",
+    /** Pop elements with the lowest scores. */
+    MIN = "MIN",
+}
+
+/**
+ * @internal
+ */
+export function createZMPop(
+    keys: string[],
+    modifier: ScoreFilter,
+    count?: number,
+): command_request.Command {
+    const args: string[] = [keys.length.toString()].concat(keys);
+    args.push(modifier);
+
+    if (count !== undefined) {
+        args.push("COUNT");
+        args.push(count.toString());
+    }
+
+    return createCommand(RequestType.ZMPop, args);
 }